constants → whitelistedModules + dynamic context
Context is built from the parent item’s scope. Items that `canEvaluate` blesses will be in the `vm` context as well as any modules that are in the `whitelistedModules` array. Entries in `whitelistedModules` should either be an absolute path to a local module (e.g. `require.resolve('./module')`) or the module name of an externally required module (e.g. `jsxstyle`).