Remove ref/unref
With this change, styles will linger in the document after the original component has unmounted.
There should be no issues with collision, though, and keeping the styles around doesn’t seem to affect performance.

This needs to be more thoroughly tested, but I ran into no issues with the Smyte production app.

A couple of outstanding TODOs:
- Find out browser’s insertRule limits and create a new style tag when we’re getting close to that limit.
- Make sure this doesn’t cause performance issues in older browsers.